+++ src/imu6500/imu_DMP6.hpp
@@ -35,6 +35,6 @@ namespace imu6500_dmp
   uint64_t get_last_baseline_reset();
   bool setupLowPowerMode();
   bool shutdown();
-  bool SetWakeOnMotionThresh(uint8_t motion_thresh_mg);
+  bool SetWakeOnMotionThresh(float motion_thresh_mg);
 
 }
